+++ src/core/component.ts
@@ -8,7 +8,8 @@ import { getOption } from "./config";
 import Configuration, { bindOptionWatchers, subscribeOnUpdates } from "./configuration";
 import { IConfigurable } from "./configuration-component";
 import { IExtension, IExtensionComponentNode } from "./extension-component";
-import { camelize, extractScopedSlots, forEachChildNode, toComparable } from "./helpers";
+import { camelize, forEachChildNode, toComparable } from "./helpers";
+import { discover as discoverTemplates } from "./templates-discovering";
 
 interface IWidgetComponent extends IConfigurable {
     $_instance: any;
@@ -28,19 +29,6 @@ interface IBaseComponent extends IVue, IWidgetComponent, IEventBusHolder {
     $_createEmitters: () => void;
     $_fillTemplate: () => void;
     $_processChildren: () => void;
-}
-
-function asConfigurable(vueComponent: IVue): IConfigurable | undefined {
-    if (!vueComponent.$vnode) {
-        return undefined;
-    }
-
-    const configurable = vueComponent.$vnode.componentOptions as any as IConfigurable;
-    if (!configurable.$_config || !configurable.$_config.name) {
-        return undefined;
-    }
-
-    return configurable;
 }
 
 const Vue = VueType.default || VueType;
@@ -119,13 +107,6 @@ const BaseComponent: VueConstructor<IBaseComponent> = Vue.extend({
         },
 
         $_getIntegrationOptions(): object {
-            const TEMPLATE_PROP = "template";
-
-            function shouldAddTemplate(child: IVue) {
-                return TEMPLATE_PROP in child.$props
-                && (child.$vnode.data && child.$vnode.data.scopedSlots);
-            }
-
             const result: Record<string, any> = {
                 integrationOptions:  {
                     watchMethod: this.$_getWatchMethod(),
@@ -133,25 +114,13 @@ const BaseComponent: VueConstructor<IBaseComponent> = Vue.extend({
                 ...this.$_getExtraIntegrationOptions(),
             };
 
-            const templates = extractScopedSlots(this.$scopedSlots, Object.keys(this.$slots));
-
-            this.$children.forEach((child: IVue) => {
-                const configurable = asConfigurable(child);
-                if (!configurable) {
-                    return;
-                }
-
-                if (shouldAddTemplate(child)) {
-                    const templateName = `${configurable.$_config.fullPath}.${TEMPLATE_PROP}`;
-                    templates[templateName] = child.$scopedSlots.default;
-                    result[templateName] = templateName;
-                }
-            });
+            const templates = discoverTemplates(this);
 
             if (Object.keys(templates).length) {
                 result.integrationOptions.templates = {};
-                Object.keys(templates).forEach((name: string) => {
-                    result.integrationOptions.templates[name] = this.$_fillTemplate(templates[name], name);
+                Object.keys(templates).forEach((templateName: string) => {
+                    result[templateName] = templateName;
+                    result.integrationOptions.templates[templateName] = this.$_fillTemplate(templates[templateName], templateName);
                 });
             }
