+++ src/core/helpers.test.ts
@@ -1,4 +1,4 @@
-import { extractScopedSlots, isEqual, toComparable } from "./helpers";
+import { isEqual, toComparable } from "./helpers";
 
 describe("toComparable", () => {
 
@@ -40,35 +40,5 @@ describe("isEqual", () => {
         const testDate1 = new Date(2018, 9, 9);
         const testDate2 = new Date(2018, 9, 9);
         expect(isEqual(testDate1, testDate2)).toBe(true);
-    });
-});
-
-describe("extractScopedSlots", () => {
-    // https://github.com/vuejs/vue/issues/9443
-
-    it("removes non-functional fields", () => {
-        const actual = extractScopedSlots(
-            {
-                a() { return true; },
-                b: 123,
-                c: true,
-                d: undefined
-            },
-            []
-        );
-
-        expect(Object.keys(actual)).toEqual(["a"]);
-    });
-
-    it("removes non-scoped slots", () => {
-        const actual = extractScopedSlots(
-            {
-                a() { return true; },
-                b() { return true; }
-            },
-            ["b"]
-        );
-
-        expect(Object.keys(actual)).toEqual(["a"]);
     });
 });
